Add BulletPointMenuButton component to CustomSkillHeader, and modify fixtures and type file
