+++ packages/@coorpacademy-components/src/organism/custom-skill-header/types.ts
@@ -1,3 +1,5 @@
+import {BulletPointMenuButtonProps} from '../../molecule/bullet-point-menu-button/types';
+
 type TagProps = {
   label: string;
   type: 'success' | 'progress' | 'warning';
@@ -7,4 +9,5 @@ export type CustomSkillHeaderProps = {
   onQuitClick: () => void;
   title: string;
   tag: TagProps;
+  bulletPointMenuButton: BulletPointMenuButtonProps;
 };
